1.4.2   Added API command to read settings
